web_app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.78.0
python-dotenv==1.0.1
python-multipart==0.0.20
